DSPY_GPT/requirements.txt
Flask
Flask-SQLAlchemy
dspy-ai
openai
yfinance
pandas
python-dotenv
Markdown
reportlab
